fix(faq): rename faq-json-ld.ts → json-ld.ts so .d.ts matches subpath
The tsup entry key `components/faq/json-ld` renames the bundle output to
`dist/components/faq/json-ld.{js,cjs}`, but tsc-emitted .d.ts mirrors the
SOURCE filename and produced `dist/components/faq/faq-json-ld.d.ts` — which
the `./components/faq/json-ld` exports entry in `package.json` does NOT
reference.

Net result: the new subpath resolved at runtime but TypeScript saw
"Module ... has no exported member 'FaqSchemaOptions'" / "buildFaqJsonLdFromFaqs"
(broke Vercel type-check on every hub consumer).

Fix: rename the source to `json-ld.ts` so the tsup bundle, the tsc
declarations, AND the package.json exports all land at the same dist path.
