+++ openframe-frontend-core/tsup.config.ts
@@ -56,8 +56,12 @@ export default defineConfig([
       // FAQ JSON-LD builder — pure (no React, no browser APIs), so Server
       // Components in consumers can import it via the './components/faq/json-ld'
       // subpath WITHOUT crossing the "use client" boundary that the sibling
-      // './components/faq' (FaqSection) subpath carries.
-      'components/faq/json-ld': 'src/components/faq/faq-json-ld.ts',
+      // './components/faq' (FaqSection) subpath carries. Source filename
+      // matches the entry key (`json-ld.ts`) so tsc-emitted `.d.ts` and
+      // tsup-emitted `.js`/`.cjs` land at the SAME `dist/components/faq/json-ld.*`
+      // paths that `package.json#exports` references — without this match the
+      // subpath resolves at runtime but TypeScript sees "no exported member".
+      'components/faq/json-ld': 'src/components/faq/json-ld.ts',
     },
     format: ['esm', 'cjs'],
     dts: false,
